mysite/requirements.txt
Django==2.2.7
django-crispy-forms==1.8.0
django-python3-ldap==0.11.2
ldap3==2.5
pyasn1==0.4.4
pyasn1-modules==0.2.7
python-ldap==3.2.0
pytz==2019.3
sqlparse==0.3.0
